+++ CHANGES.md
@@ -1,6 +1,5 @@
 
-
-- Handle cmdliner 1.1.0 deprecations.
+- `uuidtrip` support for `cmdliner` 2.0 and handle 1.1.0 deprecations.
 
 v1.0.3 2022-02-03
 -----------------